Add comments, MJPG to MP4V, requirements.txt added
Code was not working on my Windows 10 system as the camera would not display. 
I initially thought this was an issue with the codec, and I discovered that MP4V will work and left a comment as such. 
However, the problem was actually the camera index being incorrect. Added comments in two files and a very basic check. 
Could potentially be automated by cycling through values until cap.isOpened() is true, but this may not work if multiple cameras are connected. A user-selectable dropdown or some such could also possibly be added. For now, I just commented where and how to change the value. 
Added a requirements.txt to make installation a bit easier. Generated with pipreqs.